+++ src/Calendar/Controllers/TeamsController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Calendar.Data;
 using Calendar.Models;
+using Calendar.Helpers;
 
 namespace Calendar.Controllers
 {
@@ -33,6 +34,9 @@ namespace Calendar.Controllers
         // GET: Teams/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (!User.IsInRole(Constants.ROLE_ADMIN))
+                return NotFound();
+
             if (id == null)
             {
                 return NotFound();
@@ -50,6 +54,9 @@ namespace Calendar.Controllers
         // GET: Teams/Create
         public IActionResult Create()
         {
+            if (!User.IsInRole(Constants.ROLE_ADMIN))
+                return NotFound();
+
             return View();
         }
 
@@ -60,6 +67,9 @@ namespace Calendar.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("ID,CalendarStyle,Description,Name")] Team team)
         {
+            if (!User.IsInRole(Constants.ROLE_ADMIN))
+                return NotFound();
+
             if (ModelState.IsValid)
             {
                 _context.Add(team);
@@ -72,6 +82,9 @@ namespace Calendar.Controllers
         // GET: Teams/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (!User.IsInRole(Constants.ROLE_ADMIN))
+                return NotFound();
+
             if (id == null)
             {
                 return NotFound();
@@ -92,6 +105,9 @@ namespace Calendar.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("ID,CalendarStyle,Description,Name")] Team team)
         {
+            if (!User.IsInRole(Constants.ROLE_ADMIN))
+                return NotFound();
+
             if (id != team.ID)
             {
                 return NotFound();
@@ -123,6 +139,9 @@ namespace Calendar.Controllers
         // GET: Teams/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!User.IsInRole(Constants.ROLE_ADMIN))
+                return NotFound();
+
             if (id == null)
             {
                 return NotFound();
@@ -142,6 +161,9 @@ namespace Calendar.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!User.IsInRole(Constants.ROLE_ADMIN))
+                return NotFound();
+
             var team = await _context.Team.SingleOrDefaultAsync(m => m.ID == id);
             _context.Team.Remove(team);
             await _context.SaveChangesAsync();
